Extract migrateContentElement into separate file called cloneContentElement
